Modularized app a bit
