+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
-import FormSubmit from './form/form-submit'
+import FormSubmit from './components/form/form-submit'
+import Base from './components/initialSetup/base'
 import './App.css';
 
 class App extends Component {
@@ -8,13 +8,7 @@ class App extends Component {
     return (
       <div>
         <div className="App">
-          <header className="App-header">
-            <img src={logo} className="App-logo" alt="logo" />
-            <h1 className="App-title">Welcome to React</h1>
-          </header>
-          <p className="App-intro">
-            To get started, edit <code>src/App.js</code> and save to reload.
-          </p>
+          <Base />
         </div>
         <div className="App">
           <FormSubmit />
